Re-register MFA brakes OATH token activation
Whenever a user request to their respective service desk to reset the registration of MFA authentication methods from the portal, the link between the activated OATH device and the token in the portal is broken. I was able to reproduce this in my lab.
The GA will need to re-active the token, otherwise if the user uses the OATH device to get a new code, the authentication will fail with an error 500121.