Wrong descriptions for 2 steps.
In the "Update domain federation settings" where the documentation states that: "set federatedIdpMfaBehavior to rejectMfaByFederatedIdp"
We are telling the customer that "Set the Staged Rollout for Azure MFA to Off. Users will once again be redirected to your on-premises federation server for MFA."
However, the above statement would be true if instead of setting "federatedIdpMfaBehavior to rejectMfaByFederatedIdp" we configure it to "enforceMfaByFederatedIdp"

In the Rollback plan, is the other way around, we are telling the customer to change for "enforceMfaByFederatedIdp" but we are describing it as: "Users will no longer be redirected to your on-premises federation server for MFA, whether they’re targeted by the Staged Rollout tool or not. Note this can take up to 24 hours to take effect."